back-end/requirements.txt
openai
azure-ai-inference
azure-search-documents
azure-core 
python-dotenv
vite
uvicorn
fastapi
lucide-react
react-markdown